Move agent code to src/agents/
