+++ src/agents/pi-agent-extension.ts
@@ -3,7 +3,7 @@
  * @module
  */
 import type { ExtensionAPI } from "@mariozechner/pi-coding-agent";
-import { getOrOpenDb, pushEvent } from "./event-queue.ts";
+import { getOrOpenDb, pushEvent } from "../event-queue.ts";
 import { loadAgentDef } from "./file-agent-loader.ts";
 import {
   buildAgentAppendPrompt,
