+++ src/agents/pi-agent-shared.test.ts
