+++ src/agents/pi-agent-shared.ts
@@ -15,7 +15,7 @@ import {
   getClaimant,
   getEventsSince,
   pushEvent,
-} from "./event-queue.ts";
+} from "../event-queue.ts";
 import {
   loadAgentDef,
   type AgentDef,
@@ -29,9 +29,9 @@ import {
   readMemoryBlockValue,
   writeMemoryBlockValue,
   renderMemoryBlocksPrompt,
-} from "./memory/memory.ts";
-import { nextTick } from "./lib/promise.ts";
-import { renderTemplate } from "./lib/template.ts";
+} from "../memory/memory.ts";
+import { nextTick } from "../lib/promise.ts";
+import { renderTemplate } from "../lib/template.ts";
 
 /** Best-guess provider from a model ID string. */
 export const guessProvider = (model: string): string | undefined => {
