+++ src/dashboard.ts
@@ -5,7 +5,7 @@ import type {
   ExtensionContext,
   Theme,
 } from "@mariozechner/pi-coding-agent";
-import type { AgentDef } from "./file-agent-loader.ts";
+import type { AgentDef } from "./agents/file-agent-loader.ts";
 import type { Event } from "./lib/event.ts";
 import { getEventsSince } from "./event-queue.ts";
 import { getDaemonStatus } from "./pidfile.ts";
