+++ src/index.ts
@@ -25,8 +25,8 @@ import {
   registerAgentHooks,
   registerBusytownTools,
   buildAgentAppendPrompt,
-} from "./pi-agent-shared.ts";
-import { listAgentDefs, loadAgentDef } from "./file-agent-loader.ts";
+} from "./agents/pi-agent-shared.ts";
+import { listAgentDefs, loadAgentDef } from "./agents/file-agent-loader.ts";
 import { collect } from "./lib/generator.ts";
 
 const resolveDbPath = (projectRoot: string): string =>
